% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_create_cache.R
\name{add_directory_to_cache}
\alias{add_directory_to_cache}
\title{Add \code{\link{scb_list}} call to cache}
\usage{
add_directory_to_cache(cache, lang, database_id, id, depth, call_tracker)
}
\arguments{
\item{cache}{Current cache}

\item{lang}{Language: from \code{\link{scb_create_cache}}}

\item{database_id}{Database to search: from \code{\link{scb_create_cache}}}

\item{id}{Path for querying with \code{\link{scb_list}}}

\item{depth}{Current depth: 1 = top level directory}

\item{call_tracker}{Current call_tracker instance}
}
\value{
data.frame showing ID, depth, type, and name of each directory,
  subdirectory, and table, with tables also including information on
  variables contained
}
\description{
This function makes up the main body of \code{\link{scb_create_cache}}. It works by
recursively calling \code{\link{scb_list}}, each time adding the directory list and any
table metadata as appropriate.
}
\details{
For more information, see \code{\link{scb_create_cache}}.
}
