% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_search.R
\name{add_to_cache}
\alias{add_to_cache}
\title{Add scb_list call to cache}
\usage{
add_to_cache(cache, lang, database_id, id, depth, call_tracker)
}
\arguments{
\item{cache}{Current cache}

\item{lang}{Language: should be inherited}

\item{database_id}{Database to search: should be inherited}

\item{id}{Path for querying with scb_list()}

\item{depth}{How deep in the subdirectories we are}

\item{call_tracker}{Current call_tracker instance: should be created in
scb_create_cache()}
}
\value{
data.frame showing ID, type, depth, and text description of each
  directory, subdirectory, and table in the database
}
\description{
Given a list of highest level directories from scb_create_cache(), the
function goes through each directory and subdirectory in turn, cataloging
each level and storing it in memory, which is then returned.
}
