% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_time_to_year}
\alias{convert_time_to_year}
\title{Convert string time to numeric year}
\usage{
convert_time_to_year(time_value)
}
\arguments{
\item{time_value}{The time value returned from API query}
}
\value{
Integer value of year, or NA if unable to convert
}
\description{
\href{SCB:s}{www.scb.se} database stores time data in a few different formats,
many of which has no standard conversion to numeric or time objects. This
function uses the known formats to convert time data to a numeric year, to
enable easy comparisons. Months and quarters are ignored, as they are not
present in all tables.
}
\details{
The formats currently supported are listed below; each representing 2018,
January 2018, or the range of years 2018-2019: \itemize{\item "2018" \item
"2018M01" \item "2018m01" \item "2018K1" \item "2018k1" \item "2018-2019"
\item "2018/2019}
}
