% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpet_table_metadata}
\alias{interpet_table_metadata}
\title{Convert metadata to intelligible data frame}
\usage{
interpet_table_metadata(id, metadata)
}
\arguments{
\item{id}{ID of table}

\item{metadata}{Metadata returned from scb_list() call to table ID}
}
\value{
Data table containing id, name, variables, and values for table
}
\description{
Superceded by interpret_table_variables(), which provides more efficient
storage of the pertinent data contained in the metadata. Kept here in case a
new use for it is discovered, but will likely be removed in a future update.
}
