% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_create_cache.R
\name{scb_create_cache}
\alias{scb_create_cache}
\title{Create directory and table list for quick searches}
\usage{
scb_create_cache(lang = "en", database_id = "ssd")
}
\arguments{
\item{lang}{"en" English or "sv" Swedish}

\item{database_id}{Database to search}
}
\value{
data.frame showing ID, type, depth, and text description of each
  directory, subdirectory, and table in the database, to be saved in /data
}
\description{
Creates a dataset that can be saved in /data and used for quick searches.
Needed since API requests make searching slow, and the SCB API limits
requests to maximum 10 per 10 seconds, which making it even slower.
}
