% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_create_cache.R
\name{scb_create_directory_cache}
\alias{scb_create_directory_cache}
\title{Map directory and store to flat data frame}
\usage{
scb_create_directory_cache(lang = "en", database_id = "ssd")
}
\arguments{
\item{lang}{"en" English or "sv" Swedish}

\item{database_id}{Database to search}
}
\value{
data.frame showing ID, type, depth, and text description of each
  directory, subdirectory, and table in the database, to be saved in /data
}
\description{
Map out the directory structure in the given database for the given language.
This can then be saved in /data and used for quick directory searches. This
can then be used as input in table caching function.
}
