% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_create_cache.R
\name{scb_create_table_cache}
\alias{scb_create_table_cache}
\title{Store metadata on tables}
\usage{
scb_create_table_cache(lang = "en", database_id = "ssd",
  cached_directory = NULL)
}
\arguments{
\item{lang}{"en" (English) or "sv" (Swedish)}

\item{database_id}{"ssd" for standard database}

\item{cached_directory}{Created by scb_create_directory_cache: leave blank to load from /data}
}
\value{
data.table showing ID, name, and metadata for each table in database
}
\description{
Store metadata on tables
}
