% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scb_directory_cache}
\alias{scb_directory_cache}
\title{Cached view of ssd directories}
\format{A data frame with ca 2797 rows and 4 columns:
\describe{
  \item{id}{id to use to query directory or table}
  \item{type}{l for directory, t for table},
  \item{depth}{1 = highest level directory, incremented for each subdirectory}
  \item{text}{Information on database / table}
}}
\source{
rscb::scb_create_directory_cache()
}
\usage{
scb_directory_cache
}
\description{
Data frame containing information on all directories, subdirectories, and
tables in the ssd database (the main database for national statistics, and
currently the only one accessible via API).
}
\keyword{datasets}
