% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_search.R
\name{scb_list}
\alias{scb_list}
\title{List directory structure or table metadata}
\usage{
scb_list(lang = "en", database_id = "ssd", id = NULL,
  unnest_variables = TRUE, call_tracker = NULL)
}
\arguments{
\item{lang}{"en" for English, "sv" for Swedish}

\item{database_id}{Database to search}

\item{id}{Path to search in database}

\item{unnest_variables}{TRUE will unnest the variable data to be easier to
read}

\item{call_tracker}{Used in internal package functions}
}
\value{
A 2-element list, where the first element is the status_code from
  \code{\link[httr]{GET}}, and the second is a data.frame containing the
  parsed respon, or an empty data.frame if the status_code is not 200
}
\description{
By default, will return a list of all highest-level directories, in English,
in the ssd database. If an ID path is provided, either all subdirectories to
the specified path will be shown, or, if the ID path refers to a table, the
table metadata.
}
\details{
The database_id argument should normally not be touched; "ssd" is the main
database for national statistics and, as of March 2019, the only one
accessible via the API and this package. To see if any other databases are
available, one can set the database_id argument to "": i.e., call
scb_list(database_id = ""). The ID is provided as a path, e.g.
"AM/AM0101/AM0101A", where each element of the path refers to either a
directory, a subdirectory, or a table ID. The IDs can most easily be
determined by either sequentially interrogating the database (starting with
no ID argument), or by looking through the cached data in scb_cache.

The function uses the
\href{https://cran.r-project.org/web/packages/httr/index.html}{httr} package
to submit the API request, and
\href{https://cran.r-project.org/web/packages/jsonlite/index.html}{jsonlite}
to parse the response, which it then returns. If the ID path refers to a
specific table, the returned data will contain all metadata available for
that table, rather than a directory list; this metadata will be unnested
using \code{\link[tidyr]{unnest}} to make it more readable.

If the \code{\link[httr]{GET}} call returns with status code 429 (too many
requests), the function will wait until the last call to the API has cleared,
then try again. This normally resolves within, at most, 10 seconds, but there
is also a fixed limit of 50,000 attempts until the function exits to prevent
infinite loops.
}
\examples{
\dontrun{
scb_list()
scb_list(id = "AM/AM0101/AM0101A")
scb_list(lang = "sv", id = "LE")
}
}
