% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_query.R
\name{scb_query}
\alias{scb_query}
\title{Query a table}
\usage{
scb_query(table_id, ..., lang = "en", database_id = "ssd")
}
\arguments{
\item{table_id}{ID path of table to query}

\item{...}{Arguments to query table with, each like: list(code = code_to_query, filter = filter_type, values = c(values_to_filter))}

\item{lang}{Language "en" English or "sv" Swedish}

\item{database_id}{ID of database; currently only "ssd"}
}
\value{
data.frame containing response from POST query
}
\description{
Query a table
}
