% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_search.R
\name{scb_search}
\alias{scb_search}
\title{Search for directory or table in database}
\usage{
scb_search(lang = "en", database_id = "ssd", search_term)
}
\arguments{
\item{lang}{"en" English or "sv" Swedish}

\item{database_id}{Database to search}

\item{search_term}{Text to search for in directory and table names}
}
\description{
Search for directory or table in database
}
