% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_call_tracker}
\alias{update_call_tracker}
\title{Keep track of API calls}
\usage{
update_call_tracker(tracker = NULL)
}
\arguments{
\item{tracker}{The current tracking data}
}
\value{
Data frame containing currently tracked calls
}
\description{
Manages SCB limit of 10 calls per 10 seconds to ensure smooth running of
functions reliant on API calls.
}
