% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_argument}
\alias{validate_argument}
\title{Check if argument matches expectation}
\usage{
validate_argument(argument, exp_dim, exp_len, exp_type)
}
\arguments{
\item{argument}{The argument to be checked}

\item{expectation}{The expected form of the argument}
}
\value{
A string indicator: "valid" if matches, "invalid" if not, and
  "convertible" if invalid but can be converted.
}
\description{
Given an argument of any kind, the function deconstructs it and makes sure it
corresponds to the expectation. It checks for similar struture, type, and if
it appears to be convertible.
}
