% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpet_table_metadata}
\alias{interpet_table_metadata}
\title{Convert metadata to intelligible data frame}
\usage{
interpet_table_metadata(id, metadata)
}
\arguments{
\item{id}{Identifier to locate in database listing}

\item{metadata}{Metadata returned from scb_list() query with table id}
}
\value{
Data table containing id, name, variables, and values for table
}
\description{
Convert metadata to intelligible data frame
}
