% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpret_table_variables}
\alias{interpret_table_variables}
\title{Extract pertinent information from table metadata}
\usage{
interpret_table_variables(table_vars)
}
\arguments{
\item{metadata}{Output from scb_list() call with table id as argument}
}
\value{
data.frame containing pertinent data
}
\description{
Takes the results from an scb_list() call to the table id and converts to a
more intelligible format.
}
