% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_create_cache.R
\name{scb_create_cache}
\alias{scb_create_cache}
\title{Create cache of database}
\usage{
scb_create_cache(lang = "en", database_id = "ssd", initial_id = "")
}
\arguments{
\item{lang}{Supported languages: "en" English}

\item{database_id}{Supported databases: "ssd"}

\item{initial_id}{From where to start caching: default top level}
}
\value{
Cached database containing directory structure and pertinent extracts
  from table metadata
}
\description{
Creates a local and condensed copy of the given database for the given
language, with the option of starting at a particular ID path to only cache a
subset of the data.
}
\details{
The function is currently only tested with the ssd database, as that was the
only one in existence at the time of writing. However, any database that
follows the same structure should work. The output is condensed, with only
the most pertinent information being stored, in order to minimize file size.

The function works by recursively calling scb_list() and storing the results.
}
