% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_create_cache.R
\name{scb_create_cache}
\alias{scb_create_cache}
\title{New function for caching}
\usage{
scb_create_cache(lang = "en", database_id = "ssd", initial_id = "")
}
\arguments{
\item{lang}{Supported languages: "en" English}

\item{database_id}{Supported databases: "ssd"}

\item{initial_id}{From where to start caching: default top level}
}
\value{
Cached database containing directory structure and table metadata
}
\description{
Caches both directories and tables into one, and does so more efficiently.
}
