% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_search.R
\name{scb_search}
\alias{scb_search}
\title{Search for directory or table in database}
\usage{
scb_search(search_term, cached_directory = NULL, ignore_case = FALSE,
  directory_or_table = "any")
}
\arguments{
\item{search_term}{Regex to search for in directory and table names}

\item{cached_directory}{Created by scb_create_cache}

\item{ignore_case}{Instruction passed to grepl}

\item{directory_or_table}{"any", "l" (directory), or "t" (table)}
}
\description{
Currently only cached search implemented, and thus requires cached_directory,
created by scb_create_cache(), as an argument. This will be amended in the
future to provide search through uncached directory
}
