% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scb_table_cache}
\alias{scb_table_cache}
\title{Cached view of ssd tables}
\format{A data frame with x rows and 8 columns:
\describe{
  \item{id}{id to use to query table}
  \item{table_name}{name of table}
  \item{code}{variable code}
  \item{text}{description of variable}
  \item{elimination}{aggregation behaviour}
  \item{time}{indicates variable contains time data}
  \item{values}{possible values for variable}
  \item{valueTexts}{description of value}
}}
\source{
rscb::scb_create_table_cache()
}
\usage{
scb_table_cache
}
\description{
Data frame containing metadata for all tables in ssd database (the main
database for national statistics, and currently the only one accessible via
API).
}
\keyword{datasets}
