% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_scb_list}
\alias{try_scb_list}
\title{Try scb_list() and catch 429 responses}
\usage{
try_scb_list(lang, database_id, id, call_tracker)
}
\arguments{
\item{lang}{Language}

\item{database_id}{Database}

\item{id}{ID to query}

\item{call_tracker}{Current call tracker}
}
\value{
Valid response from scb_list()
}
\description{
Try scb_list() and catch 429 responses
}
