% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_call_tracker}
\alias{update_call_tracker}
\title{Manage API calls in last 10 seconds}
\usage{
update_call_tracker(tracker = NULL)
}
\arguments{
\item{tracker}{Currently tracked calls}
}
\value{
data.frame containing currently tracked calls
}
\description{
Manages list of all API calls made in the last 10 seconds. If no tracker is
provided, a new one is returned. If an already existing tracker is passed as
an argument to the function, the function will add a new call to it, then
remove all items with a timestamp > 10 seconds from current system time.
}
\details{
The SCB API limits calls to 10 per 10 seconds (rolling). If this number is
exceeded, \code{\link[httr]{GET}} returns a 429 response. We could use
\code{\link[base]{Sys.sleep}} with time set to 10 seconds when this occurs,
to reset the counter, but this would be inefficient: since the limit is
rolling, we should only wait until the oldest call is > 10 seconds in the
past, and then attempt a new query. This is what this function facilitates.
}
